Polishing the Lamented Islands
Changed out an attack for the Lamentor. Completely redone the way the Lamented Islands are generated. Other notes such as a few additions of tiles to other dungeons. Overall still need to do sounds for the Ashed King and some minor bug fixes. After that, should go back to some polishing then release